broadcastController intent to open AlarmReceiver needs to be fixed
